Add initial Docker-based CI script
