+++ Packaging/old/server_notes.txt
